+++ backend-java/src/main/java/com/rpsoft/asistencia/dtos/ReporteResponseDto.java
@@ -14,8 +14,8 @@ import java.time.LocalDateTime;
  */
 @Value
 public class ReporteResponseDto {
-    Long id;
-    Long practicanteId;
+    Integer id;
+    Integer practicanteId;
     String practicanteNombre;
     String descripcion;
     String tipo;
